+++ tests/test_settings.py
@@ -2532,7 +2532,7 @@ def main() -> None:
     defaults = {}
     setting_ids = set()
 
-    assert data.version == "2.5.0"
+    assert data.version == "2.5.1"
     assert (
         localization["quick_look_card_integration_group"]["en"]
         == "Mod Integration: Quick Look Card"